2_CNN/convolucoes/requirements.txt
glob2==0.7
opencv-contrib-python==4.4.0.42
opencv-python==4.4.0.42
pandas==1.1.1
Pillow==7.2.0
scikit-learn==0.23.2
torch==1.6.0
torchvision==0.7.0
ipynb-py-convert